requirements.txt
Bootstrap_Flask==2.4.0
Flask==3.0.3
flask_wtf==1.2.1
python-dotenv==1.0.1
WTForms==3.1.2

requests~=2.32.3
Werkzeug~=3.0.4